+++ src/main/java/com/project/lms/entity/SubjectInfoEntity.java
@@ -8,4 +8,8 @@ public class SubjectInfoEntity {
     @Id  @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "sub_seq") private Long subSeq;
     @Column(name = "sub_name") private String subName;
+
+    public String getSubName() {
+        return subName;
+    }
 }
