+++ src/main/resources/application.properties
@@ -29,6 +29,8 @@ permission.permit-all-urls[1]= /api/member/join/**
 permission.permit-all-urls[2]= /swagger/**
 permission.permit-all-urls[3]= /swagger-ui/**
 permission.permit-all-urls[4]= /v3/api-docs/**
+permission.permit-all-urls[5]= /api/score/**
+
 
 
 # redis
